server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi-pagination>=0.12.34",
    "fastapi[standard]>=0.115.11",
    "httpx>=0.28.1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pyinstaller>=6.12.0",
    "requests>=2.32.3",
    "tortoise-orm>=0.24.2",
]
